improve form > Time Picker no longer relies on absolute positioning
